pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixeldino"
version = "0.1.0"
description = "Generate pixelated dinosaurs with random clothes"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "Pillow>=9.0.0",
]

[project.scripts]
pixeldino = "pixeldino.cli:main"

[project.optional-dependencies]
dev = ["pytest"]
